fix: Another potential memory leak in the arena
I haven't tested, but it appears that we may not be properly
freeing this plugin in get_constraints. This handles if both a plugin
exists or doesn't exist. I think it's a mistake to assume, as we do in
the render callback, that if a plugin is non-nil then the game is
started. we should maybe have some other piece of state tell us that.
Then we can just lazily load the plugin and keep it around until we
switch the game out.