feat: Add test game in JS
This adds a test game written in JS. It's not really a game but just a
shell of the ABI that can be modified for writing test scenarios.

We had to enable WASI here on plugins because the JS compiler currently
requires WASI. We could perhaps, eventually, add some kind of "allows
wasi" flag on the user if we want to limit who can upload wasi enabled
games. For now we think it's okay.